Add files via upload
This is a simple, real life example on how to use Numpy to use large amounts of data. The spreadsheet has over 8000 rows of taxi data for 2016.  Using Numpy you can determine the mean speed for all 8000+ trips, the number of trips taken in February, the number of passengers that tipped more then $50/trip, and the number of dropoffs at JFK airport.   Note that February is assigned the number 2, and JFK airport is also assigned the number 2. 

Here is the original youtube video:  https://www.youtube.com/watch?v=0xwFgi1rnlc